+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Day 0 [17 Feb 2021] (± 1:30h)
+# Day 0 [17 Feb 2021] (± 0:30h)
 
 == No code day ==
 
@@ -16,7 +16,7 @@
 * Add populate-db script to load data into mongodb
 * Add environment variables strategy with .env
 
-# Day 2 [19 Feb 2021]
+# Day 2 [19 Feb 2021] (± 2:30h)
 
 * Add webpack along with react project
 * Add continuous deployment in heroku
@@ -24,3 +24,5 @@
   * For every merge on develop branch, a new deploy to development channel is made
   * Development channel: http://snappet.herokuapp.com/
 * Add and configure Material UI
+* Add state management with redux
+* Add react router (connected to redux)